+++ projects/igniteui-angular/src/lib/grids/column.component.ts
@@ -31,6 +31,7 @@ import {
     IgxDateFilteringOperand,
     IgxStringFilteringOperand } from '../data-operations/filtering-condition';
 import { IgxGridBaseComponent } from './grid-base.component';
+import { SortingStrategy } from '../data-operations/sorting-strategy';
 import { FilteringExpressionsTree } from '../data-operations/filtering-expressions-tree';
 /**
  * **Ignite UI for Angular Column** -
@@ -482,6 +483,33 @@ export class IgxColumnComponent implements AfterContentInit {
         this._filters = classRef;
     }
     /**
+     * Gets the column `sortStrategy`.
+     * ```typescript
+     * let sortStrategy = this.column.sortStrategy'
+     * ```
+     * @memberof IgxColumnComponent
+     */
+    @Input()
+    public get sortStrategy(): any {
+        return this._sortStrategy;
+    }
+    /**
+     * Sets the column `sortStrategy`.
+     * ```typescript
+     * this.column.sortStrategy = new CustomSortingStrategy().
+     *
+     * class CustomSortingStrategy extends SortingStrategy {
+     * ...
+     * }
+     * ```
+     * @memberof IgxColumnComponent
+     */
+    public set sortStrategy(classRef: any) {
+        this._sortStrategy = classRef;
+    }
+
+
+    /**
      * Gets the default minimum `width` of the column.
      * ```typescript
      * let defaultMinWidth =  this.column.defaultMinWidth;
@@ -737,6 +765,10 @@ export class IgxColumnComponent implements AfterContentInit {
      *@hidden
      */
     protected _filters = null;
+    /**
+     *@hidden
+     */
+    protected _sortStrategy = new SortingStrategy();
     /**
      *@hidden
      */